Refactor exeuction handler
